feat: add option to exclude embeddings from quantization
There is already an option to exclude a layer by its name, but for some
models this does not work, because the embedding layer name matches as
an empty string (this happens with mixtral and Llama). To avoid this and
still allow to exclude embeddings from quantization, use a new flag and
quantization option.